requirements.txt
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
Django==3.1
djangorestframework==3.12.2
idna==2.10
Pillow==7.2.0
psycopg2==2.8.6
Pygments==2.7.2
PySocks==1.7.1
pytz==2020.1
requests==2.25.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
